Moved ExplorerWidget to its own file
